[Refactor] Extract DataSource extension into a separate file
